requirements.txt
# Telegram Bot
python-telegram-bot==21.3

# AI Providers
openai==1.40.0
anthropic==0.34.0

# For local models (optional)
# ollama==0.3.0

# Environment and utilities
python-dotenv==1.0.1
pydantic==2.8.0
pydantic-settings==2.4.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1

# Scheduler for background tasks
apscheduler==3.10.4

# Date/time utilities
python-dateutil==2.9.0

# HTTP client for TON API
aiohttp==3.9.5

# QR code generation
qrcode[pil]==7.4.2
